requirements.txt
nltk
streamlit
nltk
coqui-tts